+++ kyoto/src/main/java/uk/ac/ic/kyoto/roguestates/USAgent.java
@@ -8,7 +8,7 @@ public class USAgent extends NonParticipant {
 
 	public USAgent(UUID id, String name, String ISO, double landArea,
 			double arableLandArea, double GDP, double GDPRate,
-			double emissionsTarget, long carbonOffset,
+			long emissionsTarget, long carbonOffset,
 			float availableToSpend, long carbonTraded) {
 		super(id, name, ISO, landArea, arableLandArea, GDP, GDPRate,
 				emissionsTarget, carbonOffset, availableToSpend, carbonTraded);
